runtime, flowctl: sync-now review follow-ups
- Resolve the task's type from the shard spec's task-type label rather
  than re-parsing the shard ID prefix, with taskShardZero returning the
  spec it already decoded.
- Encode the Begin->Done stopping-shortcut exclusion of
  acknowledged_count as an exhaustive Tail::resolves_on_done, and pin it
  with a unit test.
- Give the flowctl HTTP client a per-read timeout, so a stalled stream
  fails as a retryable Transport error instead of trusting heartbeats
  nothing enforced.
- Document that the Synced barrier also awaits trigger delivery.
- Move flowctl's stub-server sync-now tests to crates/flowctl/tests/,
  leaving the pure retry-classification table in-crate.
- Inline the single-caller reactor_front_door helper and drop redundant
  #[serde(default)] attributes.
