+++ go/protocols/runtime/runtime.proto
@@ -732,6 +732,10 @@ message Synced {
   // will increment it as they complete: the open transaction, if any, plus a
   // prior transaction still draining its connector acknowledgement. Never
   // more than two.
+  //
+  // A configured trigger delivers within that drain, and a debounced trigger
+  // fired while idle drains (and counts) the same way, so the barrier also
+  // awaits trigger delivery — possibly with nothing uncommitted.
   uint32 pending_count = 2;
   // Highest `CloseNow.seq` this leader has received, over all of its shards.
   // Sync-now is driven by shard zero's controller, whose sequences are
